% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCR_barcodes.R
\name{BCR_barcodes}
\alias{BCR_barcodes}
\title{Get 10x BCR Barcodes}
\usage{
BCR_barcodes(batches)
}
\arguments{
\item{batches}{Vector of batches to process.}
}
\description{
Read 10x BCR data and extract the barcodes of droplets classified as cells by Cell Ranger. Files are saved in Processing_QC as txt. Proceed to Pairing Analysis.
}
\keyword{preprocessing}
