% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_correction.R
\name{bg_correction}
\alias{bg_correction}
\title{RNA+ADT Background Noise Correction/Removal}
\usage{
bg_correction(
  raw_data,
  filter_data,
  isotype = "HSA",
  denoise_bg = TRUE,
  pool_HTO = NULL,
  project_name = NULL,
  mt_threshold = 0.2,
  prot.bg.threshold = c(1.5, 3),
  rna.bg.threshold = c(0, 2.5)
)
}
\arguments{
\item{raw_data}{CellRanger output raw_feature_bc_matrix.}

\item{filter_data}{CellRanger output sample_feature_bc_matrix.}

\item{isotype}{Name of the isotype used as ADT control. Default to HSA.}

\item{denoise_bg}{Denoising with background mean only. Default to TRUE.}

\item{pool_HTO}{Vector of HTOs in the assay to demultiplex. Default all HTOs.}

\item{project_name}{(optional) Project name used in the creation of the Seurat object.}

\item{mt_threshold}{Proportion of mitochondrial gene used as cutoff. Default 0.2.}

\item{prot.bg.threshold}{Vector of protein size thresholds to define drops as background. Default c(1.5, 3).}

\item{rna.bg.threshold}{Vector of RNA size thresholds to define drops as background. Default c(0, 2.5).}
}
\description{
This function removes background noise from ADT assays using DSB package and filters droplets based on RNA counts and mitochondrial proportion. It requires count/raw_feature_bc_matrix and count/sample_feature_bc_matrix from CellRanger outputs. Returns a Seurat object with denoised ADT counts in data slot of ADT assay.
If isotype controls were not included in the experiment, function results in not defining the technical component of each cell’s protein library. The values of the normalized matrix returned are the number of standard deviations above the expected ambient noise captured by empty droplets. It is strongly recommend to use isotype controls. However if these are not available, the background mean for each cell inferred via a per-cell Gaussian mixture model can theoretically be used alone to define the cell’s technical component, but this assumes the background mean has no expected biological variation. Isotype controls anchor the component of the background mean associated with noise.
If the normalize counts have a large range, select the correct background threshold by running bg_thresholds function and manually defining the values.
}
\keyword{bg}
