% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_thresholds.R
\name{bg_thresholds}
\alias{bg_thresholds}
\title{Define ADT Background Threshold}
\usage{
bg_thresholds(raw_data, filter_data, mt_threshold = 0.2, project_name = NULL)
}
\arguments{
\item{raw_data}{Path to CellRanger output raw_feature_bc_matrix.}

\item{filter_data}{Path to CellRanger output sample_feature_bc_matrix.}

\item{mt_threshold}{Proportion of mitochondrial gene used as cutoff. Default 0.2.}

\item{project_name}{(optional) Project name used in the creation of the Seurat object.}
}
\description{
This function returns the QC scatter plot and density plot to define correct thresholds for background drops from ADT assays. It requires count/raw_feature_bc_matrix and count/sample_feature_bc_matrix from CellRanger outputs. The plots are saved to the current directory as QC_scatter_noBgThreshold.png and QC_density.png. Define the thresholds by identifying the more dense population of background drops that is not close to 0. The density plot helps identify the RNA threshold. Normally, the background drops are located in the second pick.
}
\keyword{bg}
