% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain_call.R
\name{chain_call}
\alias{chain_call}
\title{Define Type of Chain and Frequency}
\usage{
chain_call(enclone_out)
}
\arguments{
\item{enclone_out}{Path to enclone output csv.}
}
\description{
Defining which type of chain (heavy chain, light kappa chain or light lambda chain) the call was and the frequency each chain was call for each cell. It returns a data frame.
}
\keyword{pairing}
