% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{changeo_in}
\alias{changeo_in}
\title{Create Input for Change-O}
\usage{
changeo_in(summary_path, airr_path, enclones, project_name)
}
\arguments{
\item{summary_path}{Path to 1_Summary.txt.}

\item{airr_path}{Path to vquest_airr.tsv.}

\item{enclones}{List of data frame after top_contig.}

\item{project_name}{Name of the project.}
}
\description{
Format the merged IMGT VQuest and enclone output for Change-O. Returns a list of data frames (also saved as RData), saves files to tsv/csv and FASTA to use with Change-O.
}
\keyword{clone}
