% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contig_selection.R
\name{contig_selection}
\alias{contig_selection}
\title{Contig Selection}
\usage{
contig_selection(
  batches,
  project_name,
  igm.igd = FALSE,
  vdj_type = "vdj_b",
  method = "top3",
  file = NULL
)
}
\arguments{
\item{batches}{Vector of batch or pool names.}

\item{project_name}{Name of the project.}

\item{igm.igd}{Logical. Call IgM+IgD+ instead of Doublet. Default FALSE.}

\item{vdj_type}{String. VDJ type, either "vdj_b" for BCR or "vdj_t" for TCR. Default "vdj_b".}

\item{method}{Heuristic to select best contigs. Can be "top3" for first 3 highest contigs based on number of reads, or "top2" for first 2 highest contigs based on reads for each locus. Default: "top3".}

\item{file}{IgBLAST output file path in airr format. Default: Processing_QC/singlets_contigs/project.name_singlets_contigs_igblast.tsv}
}
\description{
Determines best contigs for each cell using method. Saves fasta and annotation files for productive singlets.
}
\keyword{pairing}
