% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_CellRanger.R
\name{csv_CellRanger}
\alias{csv_CellRanger}
\title{Generate Input CSV File For Cell Ranger}
\usage{
csv_CellRanger(input_path, feature_ref_path, expect_cells, gex.ref, vdj.ref)
}
\arguments{
\item{input_path}{Path to the directory where the FastQ files are located; all files will be processed.}

\item{feature_ref_path}{Path to the feature reference file.}

\item{expect_cells}{Cells expected in batch.}
}
\description{
For each unique "sample name" in the specified directory, a new sample_name_CellRanger_Submission.csv file is created in the current directory.
}
\keyword{cellranger}
