% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_chain.R
\name{define_chain}
\alias{define_chain}
\title{Determine Chains Of enclone Output}
\usage{
define_chain(enclone_out, igm.igd = FALSE, project_name)
}
\arguments{
\item{enclone_out}{enclone output as data.frame, best if it is output from chain_call.}

\item{igm.igd}{Logical. Call IgM+IgD+ instead of Doublet. Default FALSE.}

\item{project_name}{Name of the project.}
}
\description{
Determination of chains based on highest read count. It returns a data frame.
}
\keyword{pairing}
