% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex.R
\name{demultiplex}
\alias{demultiplex}
\title{HTO Demultiplexing}
\usage{
demultiplex(
  dsb_Seurat,
  pool_HTO = NULL,
  method = "demuxmix",
  mt_threshold = 0.2,
  make_plots = TRUE,
  model = "auto",
  clusterInit = NULL
)
}
\arguments{
\item{dsb_Seurat}{Seurat object. Output from bg_correction.}

\item{pool_HTO}{Vector of HTOs in the assay. Default all HTOs.}

\item{method}{Demultiplex method, HTODemux, MULTI, demuxmix. Default demuxmix.}

\item{mt_threshold}{Proportion of mitochondrial gene used as cutoff. Default 0.2.}

\item{make_plots}{Logical. If TRUE plots are generated and saved as PNG. Defaults to TRUE.}

\item{model}{Character specifying the type of mixture model to used in demuxmix. Either "naive", "reg", "regpos", or "auto" (default).}

\item{clusterInit}{Optional list of numeric vectors to manually specify the droplet to component assignment used to initialize the EM algorithm of demuxmix. The name of each list element must match a row name of hto. The length of each element must match the number of columns of hto. Only the values 1 and 2 are allowed, where 1 belongs to the negative.}
}
\description{
This function uses the output of bg_correction in order to demultiplex and further remove empty droplets or multiplets. It returns a Seurat object.
}
\keyword{demux}
