% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metrics.R
\name{extract_metrics}
\alias{extract_metrics}
\title{Extract Metrics From CellRanger Output}
\usage{
extract_metrics(metrics_summary, batch, to_file = TRUE)
}
\arguments{
\item{metrics_summary}{Path to CellRanger output metrics_summary.csv}

\item{batch}{Batch name.}

\item{to_file}{Logical. Save metrics output as a csv file. Default to TRUE.}
}
\description{
This function allows you to extract Estimated number of cells, Mean reads per cell, Mean used reads per cell, Antibody reads usable per cell, Mean used reads per cell, Median UMI counts per cell from CellRanger output metrics_summary.csv and can save it as a csv file or print it to console. web_summary.html files will be copied to a Web_Summaries folder in the current directory.
}
\keyword{metrics}
