% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_pools.R
\name{find.cutoff}
\alias{find.cutoff}
\title{Find background threshold}
\usage{
find.cutoff(proba = 0.5, model, i = index.higher)
}
\arguments{
\item{proba}{Numeric value. Probability that the sample comes from the zero background component.}

\item{model}{List returned from normalmixEM.}

\item{i}{Integer. Index of the component that have the non-zero component, usually the second peak.}

\item{upper.cut}{Integer. Upper value to look for the root.}
}
\description{
Find the background cutoff for protein normalization with dsb. Returns a numeric value.
}
\keyword{preprocessing}
