% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CR_file.R
\name{get_CR_file}
\alias{get_CR_file}
\title{Get Cell Ranger File Path}
\usage{
get_CR_file(path, file)
}
\arguments{
\item{path}{String. Path to parent directory of file.}

\item{file}{String. Pattern or filename to retrieve path.}
}
\description{
Get path to a Cell Ranger file which can be saved in either outs or per_sample_outs directory. This command allows to use CellRanger Cirro output with proSCessoR functions.
}
\keyword{preprocessing}
