% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTOMultiHeatmap.R
\name{HTOMultiHeatmap}
\alias{HTOMultiHeatmap}
\title{HTO MULTI Heatmap}
\usage{
HTOMultiHeatmap(object, classification = "HTO_classification", assay = "HTO")
}
\arguments{
\item{object}{Seurat object after HTO demultiplexing.}

\item{classification}{Identity for the heatmap (rownames of heatmap); for example "HTO_classification" or "HTO_name". Default "HTO_classification".}

\item{assay}{Assay to use data from; for PTID-Visit-SortPop, use "PTID_HTO". Default "HTO".}
}
\description{
This function generates a heatmap of HTO expression per batch for method MULTI. It returns a ggplot object.
}
\keyword{demux}
