% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igm.igd_calling.R
\name{igm.igd_calling}
\alias{igm.igd_calling}
\title{IgM+IgD+ Calling}
\usage{
igm.igd_calling(singlets_prod_m, project_name)
}
\arguments{
\item{singlets_prod_m}{Data frame in airr format.}
}
\description{
Assign constant region as IgM.IgD to cells that have the same V,J and CDR3 (aa) instead of possibly calling them as Doublets in the pairing analysis.
}
\keyword{pairing}
