% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IgSADIE.R
\name{IgSADIE}
\alias{IgSADIE}
\title{Select Annotations from SADIE}
\usage{
IgSADIE(sadie_dir = NULL, project_name)
}
\arguments{
\item{sadie_dir}{Path to directory of IgSADIE results. Default: Processing_QC/project.name_IgBLAST_SADIE_Runs}

\item{project_name}{Name of the project.}
}
\description{
Select the annotations from SADIE runs and mark as liable the contigs that do not have complete sequences.
}
\keyword{pairing}
