% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_md.R
\name{join_md}
\alias{join_md}
\title{Join Metadata to Seurat Object}
\usage{
join_md(md_path, demux_seurat, SAC_HTO = NULL, print_res = TRUE)
}
\arguments{
\item{md_path}{Path to metadata.csv.}

\item{demux_seurat}{Output from demultiplex function.}

\item{SAC_HTO}{HTO which includes SAC as control. Default NULL.}

\item{print_res}{Logical. Make HTO (PTID_visit_sort-pop) heatmaps per batch and table of cell counts per HTO. Default TRUE.}
}
\description{
Join metadata file (must include HTO and Batch_Pool) and remove HTO from file. Saves results to dsb_sample_classifications.tsv to the current directory. Also saves and return clean singlets Seurat object.
}
\keyword{preprocessing}
