% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_Changeo-only.R
\name{merge_Changeo}
\alias{merge_Changeo}
\title{Merge Change-O Output with enclone-VQuest}
\usage{
merge_Changeo(clone_path, v_enclone_path, project_name, md_path = NULL)
}
\arguments{
\item{clone_path}{Change-O output path.}

\item{v_enclone_path}{Path to v-enclone.tsv}

\item{project_name}{Project name.}

\item{md_path}{Path to project.name_metadata_with_bc.csv; must include barcode. Default to proSCessoR's output.}
}
\description{
Merge heavy_clone-pass.tsv with VQuest-enclone data frame. File should not contain barcode overlaps. Returns a list of data frames and saves output as tsv and RData.
}
\keyword{changeo}
