% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_pools.R
\name{merge_pools}
\alias{merge_pools}
\title{Define ADT Background Threshold from Merged Pools}
\usage{
merge_pools(
  batches,
  mt_threshold = 0.2,
  project_name = NULL,
  extract_metrics = TRUE
)
}
\arguments{
\item{batches}{Vector of CellRanger output file name to be processed, i.e. "BCS2024_P2_AgNeg".}

\item{mt_threshold}{Proportion of mitochondrial gene used as cutoff. Default 0.2.}

\item{project_name}{(optional) Project name used in the creation of the Seurat object.}

\item{extract_metrics}{Logical. Save metrics to file. Default TRUE.}
}
\description{
This function returns the QC scatter plot and density plot to define correct thresholds for background drops from ADT assays. Returns a list of Seurat objects. It requires count/raw_feature_bc_matrix and count/sample_feature_bc_matrix from CellRanger outputs. The plots are saved to the current directory as QC_scatter_noBgThreshold.png. Define the thresholds by identifying the more dense population of background drops that is not close to 0. Normally, the background drops are located in the second peak.
}
\keyword{bg}
