% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_Vquest.R
\name{merge_Vquest}
\alias{merge_Vquest}
\title{Merge IMGT HighV-Quest Results with Enclone}
\usage{
merge_Vquest(summary_path, airr_path, enclones, project_name)
}
\arguments{
\item{summary_path}{Path to 1_Summary.txt.}

\item{airr_path}{Path to vquest_airr.tsv.}

\item{enclones}{List of data frame after top_contig.}

\item{project_name}{Name of project.}
}
\description{
Merge 1_Summary.txt and vquest_airr.tsv files with enclone data frame. Returns a list of data frames and saves output as tsv and RData.
}
\keyword{vquest}
