% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numPC.R
\name{numPC}
\alias{numPC}
\title{Best Number of PCs}
\usage{
numPC(obj_Seurat, pca_name)
}
\arguments{
\item{obj_Seurat}{Seurat object.}

\item{pca_name}{Name of PCA reduction to determine best number of PCs.}
}
\description{
This function runs PCA and determines the best number of principal components to use for finding neighbors, tSNE or UMAP analysis. It returns a number and plots the variance vs cumulative variance of each PC.
}
\keyword{pca}
