% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_and_clone.R
\name{pair_and_clone}
\alias{pair_and_clone}
\title{Run the VDJ Pipeline}
\usage{
pair_and_clone(
  outname,
  ref,
  db_name = "ogrdb",
  vdj = "vdj_b",
  fasta = NULL,
  igm_igd = "FALSE",
  work_dir = getwd(),
  python_site_packages
)
}
\arguments{
\item{outname}{Character. The output prefix for the run (e.g., "HVTN301_BCS2040").}

\item{ref}{Character. Absolute path to the reference directory.}

\item{db_name}{Character. Database name to use. Defaults to "ogrdb".}

\item{vdj}{Character. VDJ type. Defaults to "vdj_b".}

\item{fasta}{Character. Path to the fasta file. Defaults to Processing_QC/${outname}_singlets_prod_contig.fasta.}

\item{igm_igd}{Character. Flag for IGM/IGD contig selection. Defaults to "FALSE".}

\item{work_dir}{Character. The working directory where the raw data folders (e.g., P01) are located. Defaults to current working directory.}

\item{python_site_package}{Character. Path to python libraries where all python dependancies are installed for example "/home/mikey/.local/lib/python3.8/site-packages".}
}
\value{
Invisible integer of the system call exit status (0 means success).
}
\description{
This function acts as a wrapper to execute the bash-based pairing and clonotyping pipeline, handling module loads, fasta merging, and IgBLAST alignment.
}
