% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairing.R
\name{pairing}
\alias{pairing}
\title{B Cell Pairing Analysis Of IgBLAST Output}
\usage{
pairing(anno, project_name, singlets_prod = NULL, kl.light = TRUE)
}
\arguments{
\item{anno}{Output from run_IgSADIE as data.frame.}

\item{project_name}{Project name.}

\item{singlets_prod}{Path to singlets_prod_contig.tsv.}

\item{kl.light}{Logical. Keep cells with Kappa and Lambda chains. Default TRUE.}
}
\description{
Pairing analysis after chain determination. It returns a data frame.
}
\keyword{pairing}
