% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{preprocessing}
\alias{preprocessing}
\title{Pre-processing Of Raw Data by Project}
\usage{
preprocessing(
  seurat_list,
  isotype = FALSE,
  pool_HTO = NULL,
  method = "demuxmix",
  bg_threshold = seurat_list$bg_thresholds,
  make_plots = TRUE,
  denoise_bg = TRUE,
  project_name = NULL,
  model = "auto",
  clusterInit = NULL
)
}
\arguments{
\item{seurat_list}{Seurat list output from merge_pools function.}

\item{isotype}{Name of the isotype used as ADT control. If no control, FALSE. Default to FALSE.}

\item{pool_HTO}{Vector of HTOs in the assay to demultiplex. Default all HTOs.}

\item{method}{HTO demultiplexing method, HTODemux, MULTI or demuxmix. Default demuxmix.}

\item{make_plots}{Logical. Default TRUE}

\item{denoise_bg}{Denoising with background mean only. Only used when isotype=FALSE. Default to TRUE.}

\item{project_name}{(optional) Project name used in the creation of the Seurat object.}

\item{model}{Character specifying the type of mixture model to used in demuxmix. Either "naive", "reg", "regpos", or "auto" (default).}

\item{clusterInit}{Optional list of numeric vectors to manually specify the droplet to component assignment used to initialize the EM algorithm of demuxmix. The name of each list element must match a row name of hto. The length of each element must match the number of columns of hto. Only the values 1 and 2 are allowed, where 1 belongs to the negative.}

\item{prot_bg_threshold}{Vector of protein size thresholds to define drops as background. Default c(1.5, 3).}

\item{rna_bg_threshold}{Vector of RNA size thresholds to define drops as background. Default c(1, 2.5).}
}
\description{
Performs pre-processing of CellRanger data, including background noise removal, HTO demultiplexing and multiplets filtering. Returns a Seurat object.
}
\keyword{preprocessing}
