% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{preprocessing_byPool}
\alias{preprocessing_byPool}
\title{Pre-processing Of Raw Data by Pool}
\usage{
preprocessing_byPool(
  metrics_path,
  raw_path,
  filter_path,
  batch,
  isotype = "HSA",
  pool_HTO = NULL,
  method = "HTODemux",
  mt_threshold = 0.2,
  prot.bg.threshold = c(1.5, 3),
  rna.bg.threshold = c(0, 2.5),
  make_plots = TRUE,
  denoise_bg = FALSE,
  project_name = NULL,
  to_file = TRUE
)
}
\arguments{
\item{metrics_path}{Path to CellRanger metrics_summary.}

\item{raw_path}{Path to CellRanger outs/multi/count/raw_feature_bc_matrix.}

\item{filter_path}{Path to CellRanger outs/per_sample_outs/SampleName/count/sample_feature_bc_matrix.}

\item{isotype}{Name of the isotype used as ADT control. Default to HSA. If no control, FALSE.}

\item{pool_HTO}{Vector of HTOs in the assay to demultiplex. Default all HTOs.}

\item{method}{Demultiplex method, HTODemux or MULTI. Default HTODemux.}

\item{mt_threshold}{Proportion of mitochondrial genes use as cutoff. Default 0.2.}

\item{prot.bg.threshold}{Vector of protein size thresholds to define drops as background. Default c(1.5, 3).}

\item{rna.bg.threshold}{Vector of RNA size thresholds to define drops as background. Default c(0, 2.5).}

\item{make_plots}{Logical. Default TRUE}

\item{denoise_bg}{Denoising with background mean only. Default to FALSE. Only used when isotype=FALSE.}

\item{project_name}{(optional) Project name used in the creation of the Seurat object.}

\item{to_file}{Logical. Save metrics to file. Default TRUE}
}
\description{
Performs pre-processing of CellRanger data, including background noise removal, HTO demultiplexing and multiplets filtering. Returns a Seurat object.
}
\keyword{preprocessing}
