% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_NA.R
\name{replace_NA}
\alias{replace_NA}
\title{Replaces NA to No or 0}
\usage{
replace_NA(enclone_out)
}
\arguments{
\item{enclone_out}{Data frame after top_contig.}
}
\description{
Replaces NA to "No" in Single_HC and Light_Chain, and NA to 0 in Paired. It returns a data frame.
}
\keyword{pairing}
