% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna_correction.R
\name{rna_correction}
\alias{rna_correction}
\title{RNA Background Noise Correction/Removal}
\usage{
rna_correction(
  raw_data,
  filter_data,
  pool_HTO = NULL,
  project_name = NULL,
  mt_threshold = 0.2,
  hto_exp = TRUE
)
}
\arguments{
\item{raw_data}{CellRanger output raw_feature_bc_matrix.}

\item{filter_data}{CellRanger output sample_feature_bc_matrix.}

\item{pool_HTO}{Vector of HTOs in the assay to demultiplex. Default all HTOs.}

\item{project_name}{(optional) Project name used in the creation of the Seurat object.}

\item{mt_threshold}{Proportion of mitochondrial gene used as cutoff. Default 0.2.}

\item{hto_exp}{Logical. False if HTO expression is not included. Default to TRUE.}
}
\description{
This filters droplets based on RNA counts and mitochondrial proportion. It requires count/raw_feature_bc_matrix and count/sample_feature_bc_matrix from CellRanger outputs. Returns a Seurat object with denoised ADT counts in data slot of ADT assay.
}
\keyword{bg}
