% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_changeo.R
\name{run_changeo}
\alias{run_changeo}
\title{Run Change-O In R}
\usage{
run_changeo(heavy_path, light_path, project_name)
}
\arguments{
\item{heavy_path}{Path to heavy_productive-T_sc-pass.tsv.}

\item{light_path}{Path to light_productive-T_sc-pass.tsv.}

\item{project_name}{Name of the project.}
}
\description{
The function runs the clonotyping analysis after successful generation of productive-T_sc-pass.tsv files, but failing to calculate threshold. Specifically, it calculates the distance to nearest neighbor, finds the threshold and groups clones. Saves the final output to heavy and light clone-pass.tsv to current directory. Immcantation libraries must be installed since Docker is not used.
}
\keyword{clones}
