% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_enclone.R
\name{run_enclone}
\alias{run_enclone}
\title{Run enclone}
\usage{
run_enclone(batches, project_name, classifications_path = NULL)
}
\arguments{
\item{batches}{Directory names of each batch to run.}

\item{project_name}{Name of the project.}

\item{classifications_path}{Path to dsb_sample_classifications.tsv. Default "./project_name_dsb_sample_classifications.tsv".}
}
\description{
Run enclone from files listed in vector batches. The function calls enclone from the command line. Saves merged_enclone_output.csv to current directory.
}
\keyword{enclone}
