% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_ADTs.R
\name{scatter_ADTs}
\alias{scatter_ADTs}
\title{Scatter Plots of ADTs}
\usage{
scatter_ADTs(obj_Seurat, no_ADTs = NULL, project_name = NULL)
}
\arguments{
\item{obj_Seurat}{Seurat object with ADT assay.}

\item{no_ADTs}{Vector of ADTs to exclude from the graphs. Default will return graphs for all ADTs.}

\item{project_name}{Name of project. Default will name the directory as "ADT-scatterplot-default".}
}
\description{
This function creates scatter plots for each combination of ADTs. It returns a directory containing all the plots.
}
\keyword{plots}
