% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_PCA.R
\name{seurat_PCA}
\alias{seurat_PCA}
\title{Perform PCA}
\usage{
seurat_PCA(
  obj_Seurat,
  assay,
  norm_method = "sct",
  vars_regress = "percent.mt",
  pca_name = NULL
)
}
\arguments{
\item{obj_Seurat}{Seurat object.}

\item{assay}{Assay to perform the analysis on.}

\item{norm_method}{Normalization method ("sct" or "lognorm"). Defaults to "sct".}

\item{vars_regress}{Variables to regress with SCTransform; variable from obj_Seurat metadata. Defaults to mitochondrial genes.}

\item{pca_name}{Name for PCA results stored in the Seurat object. Default: If RNA assay, pca_name="pca".}
}
\description{
This function normalizes, finds variable features and scales data using SCTransform or NormalizeData from Seurat, and runs PCA. It returns a Seurat object.
}
\keyword{pca}
