% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlets_only.R
\name{singlets_only}
\alias{singlets_only}
\title{Filter Singlets}
\usage{
singlets_only(demux_Seurat)
}
\arguments{
\item{demux_Seurat}{Demultiplexed Seurat object. Output from demultiplex.}
}
\description{
This function uses the output of demultiplex in order to filter out empty drops and doublets, keeping only singlets. It returns a Seurat object.
}
\keyword{singlets}
