% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryQC.R
\name{SummaryQC}
\alias{SummaryQC}
\title{QC Summary Plots}
\usage{
SummaryQC(batches)
}
\arguments{
\item{batches}{Vector of batches with batch_ReadMetrics.csv files.}
}
\description{
This function generates QC summary plots (all batches) from CellRanger metrics reports and singlets/doublets count distributions.
}
\keyword{QC}
