% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairing.R
\name{tcr.pairing}
\alias{tcr.pairing}
\title{T Cell Pairing Analysis Of IgBLAST Output}
\usage{
tcr.pairing(anno, project_name, singlets_prod = NULL)
}
\arguments{
\item{anno}{Output from run_IgSADIE as data.frame.}

\item{project_name}{Project name.}

\item{singlets_prod}{Path to singlets_prod_contig.tsv.}
}
\description{
Pairing analysis after chain determination. It returns a data frame.
}
\keyword{pairing}
