% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_contig.R
\name{top_contig}
\alias{top_contig}
\title{Create FASTA Of Top Contig From enclone Output}
\usage{
top_contig(enclone_out, project_name)
}
\arguments{
\item{enclone_out}{enclone output as data.frame, output from pairing.}

\item{project_name}{Name of project.}
}
\description{
Select top contig after pairing and generate FASTA file for all chains. It returns a FASTA file and a list of data frames.
}
\keyword{pairing}
