% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_cluster.R
\name{umap_cluster}
\alias{umap_cluster}
\title{Clustering and UMAP}
\usage{
umap_cluster(
  pca_Seurat,
  assay,
  pca_name,
  umap_name,
  resolution = 0.2,
  plots = TRUE
)
}
\arguments{
\item{pca_Seurat}{Seurat object with PCA reduction data.}

\item{assay}{Assay to perform analysis on.}

\item{pca_name}{Name of PCA data reduction to use in analysis.}

\item{umap_name}{Name of UMAP data reduction to be stored in the Seurat object.}

\item{resolution}{Resolution setting. Higher number is used for larger data sets (more clusters). Default 0.2.}

\item{plots}{Logical. Return UMAP clustering saved as PDF. Default to TRUE.}
}
\description{
This function uses the output of seurat_PCA. It returns a Seurat object and can plot the clustering in UMAP.
}
\keyword{umap}
