% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_enclone_to_changeo.R
\name{v_enclone_to_changeo}
\alias{v_enclone_to_changeo}
\title{Format V-quest/Enclone output for Change-O}
\usage{
v_enclone_to_changeo(v_enclones, project_name)
}
\arguments{
\item{v_enclones}{List of data frames from merge_Vquest.}

\item{project_name}{Name of the project.}
}
\description{
Create CellRanger-like FASTA and contig annotation files for use with Change-O. File should not contain barcode overlaps. It returns FASTA file and csv.
}
\keyword{changeo}
